Beta site release
The projects page is half done so the navigation for it has been disabled.

The CV is not yet linked to the about page so the content showing that has also been hidden.

Some further tweaks are going to be made but these are minor and will not affect the core workings of the site.